OPAL-2097 - Created a non-resource class that creates and removes packages.
